✨ Feat: 피드 뷰 컴포넌트화
